+++ daq_ini.m
@@ -19,8 +19,8 @@ s.IsContinuous = true;
 if Recmode == 1
     % for iRecHS2
     %(1):Eye_Vertical, (2):Eye_Horizontal, (3):photo sensor, (4):Trigger pulse
-    %(5):Refrect_Vertical, (6):Refrect_Horizontal
-    InCh = addAnalogInputChannel(s, dev.ID, 0:3, 'Voltage');
+    %(5):PupilSize
+    InCh = addAnalogInputChannel(s, dev.ID, 0:4, 'Voltage');
     
     % default SingleEnded, -> Differential (default)
     InCh(1).TerminalConfig = 'Differential';
@@ -29,7 +29,8 @@ if Recmode == 1
     % Channle4 is used for hardware timing.
     InCh(4).TerminalConfig = 'Differential';
     
-    %InCh(5).TerminalConfig = 'Differential';
+    %pupilsize
+    InCh(5).TerminalConfig = 'Differential';
     %InCh(6).TerminalConfig = 'Differential';
     
 elseif Recmode == 2